# 25% throughput improvement