@mcollina 2021-12-17
This was the flamegraph we did that found the problem:

![Uploading Screenshot 2021-12-17 at 17.28.27.png…]()